adjust header and IDL types
This is a combination of multiple smaller changesets that fit together in
theme. Their component changelogs follow. It should be noted that the right
way to do these types is documented in manpages-posix-dev, so going forward
there's no good reason for fuckery like the previous.

<stdlib.h> now includes stddef, limits, math, and sys/types; and <sys/types.h>
gets size_t from stddef thru a single-import macro. Now size_t will always be
defined by compiler headers to the sizeof() type.

uid_t and gid_t are now found in IDL land as Posix::id_t, which is signed. The
previous definition was unsigned, violating POSIX.

idtype_t was changed to an enum, as POSIX requires. pid_t was changed to a
signed integer.

useconds_t was corrected to be a half-phantom type just for usleep(3).

proc, vm, io, and fs were updated to use types from posixlike.idl instead of
homegrown ad-hoc ones. Advantage being that the former comes out compatible to
those from <sys/types.h> and <stdlib.h>.

+++ root/main.c
@@ -1407,7 +1407,7 @@ static void parse_initrd_args(struct htable *dest)
 /* TODO: move into bootcon.c */
 
 static int bootcon_read(
-	int32_t cookie, unsigned length,
+	int32_t cookie, off_t length,
 	uint8_t *buf, unsigned *len_p)
 {
 	*len_p = 0;
